requirements.txt
pillow
requests
numpy

# optional
torch